pharma_deal_django/requirements.txt
asgiref==3.4.1
astroid==2.3.3
autopep8==1.5.7
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
colorama==0.4.3
cryptography==3.4.7
Deprecated==1.2.12
Django==3.2.5
gcloud==0.18.3
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.19.1
idna==3.2
isort==4.3.21
jwcrypto==0.9.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauth2client==4.1.3
pipenv==2021.5.29
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.10.1
pylint==2.4.4
pyparsing==2.4.7
Pyrebase4==4.5.0
python-jwt==3.3.0
pytz==2019.3
requests==2.26.0
requests-toolbelt==0.9.1
rsa==4.7.2
six==1.14.0
sqlparse==0.3.0
toml==0.10.2
urllib3==1.26.6
virtualenv==16.7.7
virtualenv-clone==0.5.4
whitenoise==5.3.0
wrapt==1.11.2
